separate states enum
