+++ src/lib/states.ts
@@ -1,0 +1,9 @@
+export enum States {
+    AddingZoneCanvas = 1,
+    AddingZoneMap,
+    Waiting,
+    EditingZone,
+    DeletingZoneCanvas,
+    DeletingZoneMap,
+    PickPolygon
+}
